Move control-plane related types into deps/control-plane
